requirements.txt
twitchio>=2.6.0,<3.0.0
python-dotenv>=1.0.0
httpx>=0.27.0
google-genai>=1.0.0